+++ FE/Algo/app/src/main/java/com/d204/algo/data/model/Status.kt
@@ -6,8 +6,8 @@ import kotlinx.parcelize.Parcelize
 
 @Parcelize
 data class Status(
-    override val id: Int,
-    val userId: Int,
+    override val id: Long,
+    val userId: Long,
     var wisdom: Int,
     var vitality: Int,
     var strength: Int,
