+++ FE/Algo/app/src/main/java/com/d204/algo/data/repository/RankingRepository.kt
@@ -7,5 +7,5 @@ import kotlinx.coroutines.flow.Flow
 interface RankingRepository {
     suspend fun getRankings(): Flow<NetworkResult<List<Ranking>>>
     suspend fun getRankingsByTier(tier: Int): Flow<NetworkResult<List<Ranking>>>
-    suspend fun getRanking(userId: Int): Flow<NetworkResult<Ranking>>
+    suspend fun getRanking(userId: Long): Flow<NetworkResult<Ranking>>
 }
