+++ FE/Algo/app/src/main/java/com/d204/algo/data/repository/datasource/StatusDataSource.kt
@@ -4,7 +4,7 @@ import com.d204.algo.data.model.Status
 
 interface StatusDataSource {
     // remote
-    suspend fun getStatus(userId: Int): Status
+    suspend fun getStatus(userId: Long): Status
 
     // cache
