+++ FE/Algo/app/src/main/java/com/d204/algo/data/repository/remote/ProblemRemote.kt
@@ -4,6 +4,6 @@ import com.d204.algo.data.model.Problem
 
 interface ProblemRemote {
     suspend fun getProblems(): List<Problem>
-    suspend fun getProblem(problemId: Int): Problem
+    suspend fun getProblem(problemId: Long): Problem
     suspend fun isRemote(): Boolean
 }
