+++ FE/Algo/app/src/main/java/com/d204/algo/data/repository/remote/RankingRemote.kt
@@ -5,6 +5,6 @@ import com.d204.algo.data.model.Ranking
 interface RankingRemote {
     suspend fun getRankings(): List<Ranking>
     suspend fun getRankingsByTier(tier: Int): List<Ranking>
-    suspend fun getRanking(userId: Int): Ranking
+    suspend fun getRanking(userId: Long): Ranking
     suspend fun isRemote(): Boolean
 }
