+++ FE/Algo/app/src/main/java/com/d204/algo/data/source/cache/StatusCacheDataSource.kt
@@ -8,7 +8,7 @@ import javax.inject.Inject
 class StatusCacheDataSource @Inject constructor(
     private val StatusCache: StatusCache
 ) : StatusDataSource {
-    override suspend fun getStatus(userId: Int): Status {
+    override suspend fun getStatus(userId: Long): Status {
         throw UnsupportedOperationException("getStatuss is not supported for StatusCacheDataSource.")
     }
