+++ FE/Algo/app/src/main/java/com/d204/algo/remote/model/ProblemLikeModel.kt
@@ -3,9 +3,9 @@ package com.d204.algo.remote.model
 import com.squareup.moshi.Json
 
 data class ProblemLikeModel(
-    @field:Json(name = "problem_like_id") val id: Int,
-    @field:Json(name = "user_id") val userId: Int,
-    @field:Json(name = "problem_id") val problemId: Int,
+    @field:Json(name = "problem_like_id") val id: Long,
+    @field:Json(name = "user_id") val userId: Long,
+    @field:Json(name = "problem_id") val problemId: Long,
     @field:Json(name = "like_type") val likeType: Boolean,
     @field:Json(name = "memo") val memo: String,
 )
